calc_fact() を polldb に移動。a_trend, g_trend で共有。
